fixed the name conflict in the Signin page
