+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom"
 import './App.css'
-import Signin from './pages/Login'
+import Signin from './pages/Signin'
 import { Signup } from "./pages/Signup"
 import { EmailVerification } from "./pages/EmailVerification"
 import { Homepage } from "./pages/Homepage"
@@ -15,7 +15,7 @@ function App() {
         <Routes>
           <Route index element={<Signup />}/>
           <Route path="SignUp" element={<Signup/>}/>
-          <Route path="Login" element={<Signin/>}/>
+          <Route path="Signin" element={<Signin/>}/>
           <Route path="EmailVerification" element={<EmailVerification/>}/>
           <Route path="Homepage" element={<Homepage/>}/>
           <Route path="Onboarding" element={<Onboarding/>}/>
